==> notebooks/1.0.combine_json_files.ipynb <==
import os
os.chdir('..') # this resolves ImportError: attempted relative import with no known parent package
import json
import glob
from datetime import datetime as dt

def combine_all_json_files(folder_path):
    """Combines multiple JSON files into one file whilst maintaining the same
    file and JSON structure.
    
    Parameters
    -----------
    folder_path: str
        String of the path to the folder that contains the json files you want to combine 
        into one json file.
    
    
    Returns
    --------
    'all_json_data.json': json file
        File with all the json data contained in the individual json files in the folder path.
    
    
    Notes
    -------

    """
    
    input_path = folder_path + "/*.json"
    output_path = "data/processed"
    
    # create empty list - this will be populated with all the json files that we want to combine
    data = []
    
    # loop through the files in the foler and append them to the list
    for file in glob.glob(input_path):
        with open(file, "r") as f:
            data.append(json.loads(f.read()))
            
            
    # create a new file with all the json data and save in the output folder path
    with open(f"{output_path}/all_json_data.json", "w") as f:
        json.dump(data, f, indent=4)
        
    return

# set path to load files from
path = "data/raw/JSON_response"

# load all the json files into one json file
combine_all_json_files(folder_path=path)
