==> notebooks/3.0.clean_dataframe_functions.ipynb <==
import pandas as pd
import datetime
from datetime import datetime as dt
import re

raw_data = pd.read_csv("../data/processed/json_data_df.csv")
data = raw_data.copy()
df = pd.DataFrame(data)
df.head()

# Remove prefixes

def remove_prefixes(column_name):
    """Removes the prefixes e.g., 'snippet' and 'statistics' from the column titles.
    
    
    Parameters
    -----------
    column_name (str):
        A string representing a column name.
        
    Returns
    --------
    column_name (str):
        Returns the input without the prefixes specified in the prefixes_and_replacement
        dictionary.
        
    Example Use
    ------------
    df.columns = df.columns.map(remove_prefixes)
    
    Notes
    ------
    
    """
    # define the prefixes we want to remove and what we'll replace them with (nothing)
    prefixes_and_replacement = {"snippet.": "", 
                               "statistics.": "",
                               "contentDetails.": ""}
    
    for prefix, replacement in prefixes_and_replacement.items():
        if column_name.startswith(prefix):
            column_name = column_name.replace(prefix, replacement)
            
    return column_name

df.columns = df.columns.map(remove_prefixes)
df.columns

# Transform duration

def duration_to_hhmmss(duration):
    """Takes in a string of YouTube's API duration 'response' and uses a regex to return two values:
        - the duration in sconds as a timedelta object
        - the duration in hh:mm:ss format.
    
    Parameters
    ----------
    duration (str):
        YouTube's API duration response which provides the duration of a 
        video as a string with the format "PT1H21M36S" for 1 hour, 21 minutes
        and 36 seconds.
        
        
    Returns
    --------
    duration_timedelta (timedelta):
        The input transformed into a timedelta objct which provids the duration in seconds.
        
    duration_string (str):
        The input transformed into a string with format: "hh:mm:ss".
        
        
    Example Use
    ------------
    Extracting duration of YouTube video:
        - Create a column that can compare the durations of the videos with:
            
                df['duration_seconds'] = df['duration'].apply(lambda x: duration_to_hhmmss(x)[0])
                
        - Create a quickly interpretable duration column with:
        
                df['duration_string'] = df['duration'].apply(lambda x: duration_to_hhmmss(x)[1])
    
    """
    
    # set duration pattern to capture the YouTube API response which gives duration in format PT1H52M21S
    # for 1 hour, 52 minutes, 21 seconds
    duration_pattern = r"[A-Za-z]+(?:(\d+)H)?(?:(\d+)M)?(?:(\d+)S)?"
    
    # use regex to compile the pattern
    duration_regex = re.compile(duration_pattern)
    
    # match the pattern and store the groups returned in a variable
    duration_groups = duration_regex.match(duration)
    
    # store each group in a variable
    full_response, hours, minutes, seconds = duration_groups[0], duration_groups[1], duration_groups[2], duration_groups[3]
    
    # ensure we have a non-Null value for hours, minutes, and seconds, and zero pad them
    if hours != None:
        if len(hours) == 1:
            hours = '0' + hours
            
    elif hours == None:
        hours = '00'
            
    if minutes != None:
        if len(minutes) == 1:
            minutes = '0' + minutes
                       
    elif minutes == None:
        minutes = '00'
    
    if seconds != None:
        if len(seconds) == 1:
            seconds = '0' + seconds
                          
    elif seconds == None:
        seconds = '00'
     
    
    # create datetime object
    duration_string = f"{hours}:{minutes}:{seconds}"
    
#     duration_formatted = dt.strptime(duration_string, "%H:%M:%S")
    
    # use the groups retrieved to create a time object
    duration_timedelta = dt.timedelta(days=0, hours=int(hours), minutes=int(minutes), seconds=int(seconds))
    
    return duration_timedelta, duration_string

df['duration_timedelta'] = df['duration'].apply(lambda x: duration_to_hhmmss(x)[0])
df['duration_hhmmss'] = df['duration'].apply(lambda x: duration_to_hhmmss(x)[1])
df.head()

# Drop columns

def drop_columns(df):
    """Drops columns and returns a dataframe.
    
    Parameters
    -----------
    df (dataframe):
        pandas dataframe. 
        
    Returns
    --------
    df_ (dataframe)
        dataframe with columns dropped.
    
    
    
    """
    
    # create a copy of the dataframe
    df_ = df.copy()
    
    # store the columns we want to drop in a list that we'll loop through later
    cols_to_drop = ['thumbnails.default.url', 'thumbnails.default.width',
                    'thumbnails.default.height', 'thumbnails.medium.url',
                    'thumbnails.medium.width', 'thumbnails.medium.height',
                    'thumbnails.high.url', 'thumbnails.high.width',
                    'thumbnails.high.height', 'thumbnails.standard.url',
                    'thumbnails.standard.width', 'thumbnails.standard.height',
                    'thumbnails.maxres.url', 'thumbnails.maxres.width',
                    'thumbnails.maxres.height', 'localized.title', 'localized.description', 
                    'defaultAudioLanguage', 'categoryId', 'liveBroadcastContent', 
                    'dimension', 'duration', 'definition', 'caption','licensedContent', 
                    'projection', 'regionRestriction.blocked', 'contentRating.ytRating']
    

    
    # loop through columns to drop list and drop if the column exists
    for column in cols_to_drop:
        try:
            df_.drop(column, axis=1, inplace=True)
            
        except KeyError as ke:
            print(ke)
            continue
    
    return df_

df = drop_columns(df)
df.head()

#  define function that splits publishedAt into publishedAtYear, Month and Day

# Export to csv

df.to_csv("../data/processed/json_data_processed.csv", index=False )
