==> notebooks/4.4.feature_engineering_video_counts.ipynb <==
import os
os.chdir('..') # this resolves ImportError: attempted relative import with no known parent package
# import sys
# parent_dir = os.getcwd()
# sys.path.append(parent_dir)

import pandas as pd
from src.processing.feature_engineering import create_year_month_columns

raw_data = pd.read_csv("data/processed/json_data_processed.csv")
data = raw_data.copy()
df = pd.DataFrame(data)
df.head()

df = create_year_month_columns(df)
df.head()

# define a function that counts the number of videos released in a calendar month
def create_video_counts_columns(df):
    """Calculates the number of videos per month and per year based on the date in the 'publishedAt' column.
    
    Parameters
    -----------
    df : pandas dataframe
        Dataframe with 'publishedAt' column.
        
    Returns
    -------
    df : pandas dataframe
        The input dataframe with additional 'videoCountMonth' and 'videoCountYear' columns.
        
    Notes
    ------
    
    
    """
    # create copy of input dataframe to avoid any inplace changes
    df_ = df.copy()
    
    # # convert publishedAt values from string to datetime
    # df_['publishedAt'] = pd.to_datetime(df_['publishedAt'])
    
    # # use strftime to format the datetime value into just year and month so we can group by year and month
    # df_['publishedAtYear'] = df_['publishedAt'].dt.strftime("%Y")
    # df_['publishedAtMonth'] = df_['publishedAt'].dt.strftime("%m")
    # df_['publishedAtYearMonth'] = df_['publishedAt'].dt.strftime("%Y-%m")
    
    # create video count per month dataframe
    video_count_month = df_['publishedAtYearMonth'].value_counts().reset_index()
    
    # change column titles
    video_count_month.columns = ['yearMonth', 'videoCountMonth']
    
    # convert year_month to datetime for easy sorting by date
    video_count_month['yearMonth'] = pd.to_datetime(video_count_month['yearMonth'], format="%Y-%m")
    
    # add year and month columns for easy merging later
    video_count_month['year'] = video_count_month['yearMonth'].dt.strftime("%Y")
    video_count_month['month'] = video_count_month['yearMonth'].dt.strftime("%m")
    video_count_month['yearMonth'] = video_count_month['yearMonth'].dt.strftime("%Y-%m")
    
    # sort values by date
    video_count_month.sort_values(by=["yearMonth"], inplace=True)
    video_count_month.reset_index(inplace=True, drop=True)
    
    # reorder columns
    video_count_month = video_count_month[["yearMonth", "year", "month", "videoCountMonth"]]
    

    # get video count per year
    video_count_year = df_['publishedAtYear'].value_counts().reset_index()
    
    # change column titles
    video_count_year.columns = ['year', 'videoCountYear']
    
    # sort values by year
    video_count_year.sort_values(by=["year"], inplace=True)
    video_count_year.reset_index(inplace=True, drop=True)

    # merge video_count_month and video_count_year dataframes
    video_counts = pd.merge(video_count_month, video_count_year, left_on="year", right_on="year", how='left')

    # add video count data to original dataframe
    df_ = pd.merge(df_, video_counts, left_on=['publishedAtYearMonth'], right_on=['yearMonth'], how='left')

    # drop unnecessary columns (the columns in these data is contained in publishedAtYear, publishedAtMonth etc.
    df_.drop(labels=["yearMonth", "year", "month"], axis=1, inplace=True)

    # sort values and reset index
    df_.sort_values(by=['publishedAt'], inplace=True)
    df_.reset_index(drop=True, inplace=True)

    df_ = df_[['videoID', 'publishedAt', 'publishedAtYear', 'publishedAtMonth', 'publishedAtYearMonth', 
               'channelTitle', 'channelId',  'title', 'description', 'duration_timedelta', 'duration_hhmmss', 'tags', 
               'viewCount', 'likeCount', 'favoriteCount','commentCount',
               'videoCountMonth', 'videoCountYear']]


    return df_
    
    





df = create_video_counts_columns(df)
df.head()

def create_video_counts_columns(df):
    """Calculates the number of videos per month and per year based on the date in the 'publishedAt' column.
    
    Parameters
    -----------
    df : pandas dataframe
        Dataframe with 'publishedAt' column.
        
    Returns
    -------
    df : pandas dataframe
        The input dataframe with additional 'videoCountMonth' and 'videoCountYear' columns.
        
    Notes
    ------
    
    
    """
    # create copy of input dataframe to avoid any inplace changes
    df_ = df.copy()
    
    # create video count per month dataframe
    video_count_month = df_['publishedAtYearMonth'].value_counts().reset_index()
    
    # change column titles
    video_count_month.columns = ['yearMonth', 'videoCountMonth']
    
    # convert year_month to datetime for easy sorting by date
    video_count_month['yearMonth'] = pd.to_datetime(video_count_month['yearMonth'], format="%Y-%m")
    
    # add year and month columns for easy merging later
    video_count_month['year'] = video_count_month['yearMonth'].dt.strftime("%Y")
    video_count_month['month'] = video_count_month['yearMonth'].dt.strftime("%m")
    video_count_month['yearMonth'] = video_count_month['yearMonth'].dt.strftime("%Y-%m")
    
    # sort values by date
    video_count_month.sort_values(by=["yearMonth"], inplace=True)
    video_count_month.reset_index(inplace=True, drop=True)
    
    # reorder columns
    video_count_month = video_count_month[["yearMonth", "year", "month", "videoCountMonth"]]
    

    # get video count per year
    video_count_year = df_['publishedAtYear'].value_counts().reset_index()
    
    # change column titles
    video_count_year.columns = ['year', 'videoCountYear']
    
    # sort values by year
    video_count_year.sort_values(by=["year"], inplace=True)
    video_count_year.reset_index(inplace=True, drop=True)

    # merge video_count_month and video_count_year dataframes
    video_counts = pd.merge(video_count_month, video_count_year, left_on="year", right_on="year", how='left')

    # add video count data to original dataframe
    df_ = pd.merge(df_, video_counts, left_on=['publishedAtYearMonth'], right_on=['yearMonth'], how='left')

    # drop unnecessary columns (the columns in these data is contained in publishedAtYear, publishedAtMonth etc.
    df_.drop(labels=["yearMonth", "year", "month"], axis=1, inplace=True)

    # sort values and reset index
    df_.sort_values(by=['publishedAt'], inplace=True)
    df_.reset_index(drop=True, inplace=True)

    try:
        df_ = df_[['videoID', 'publishedAt', 'publishedAtYear', 'publishedAtMonth', 'publishedAtYearMonth',
                   'channelTitle', 'channelId',  'title', 'description', 'duration_timedelta', 'duration_hhmmss', 'tags',
                   'viewCount', 'likeCount', 'favoriteCount','commentCount',
                   'videoCountMonth', 'videoCountYear']]

    except KeyError as ke:
        # store error message (format is "KeyError: ['column_x', 'column_y', ... 'column_n'] not in index")
        error_message = str(ke)     
        
        # Extract the string between square brackets
        missing_columns_str = error_message.split("'")[1]
        
        # Split the string between the square brackets into individual column names
        missing_columns_list = missing_columns_str.split(", ")
        
        # Print the list of missing columns
        print(f"Warning: {ke}\nDataframe created without these columns.")

        # store remaining columns
        remaining_columns = [col for col in df_.columns if col not in missing_columns_list]
        
        # create dataframe with just remaining columns
        df_ = df_[remaining_columns]


    return df_

df = create_video_counts_columns(df)
df.head()

df.columns
